Move the three plain coordinate helpers out of the numba module
`is_valid_lat`, `is_valid_lng` and `coord2int` stopped being `njit` functions
in the previous commit but stayed in `utils_numba.py`, which is the module the
package compiles - a reader reaching for a function there is entitled to
assume it is compiled, and an uncompiled one sitting among them invites the
decorator back. They move to `timezonefinder/utils.py`, which already
re-exported all three under the same names, so the aliasing indirection goes
with them and no public name changes.

`int2coord` stays in the numba module and stays compiled: `convert2coords` and
`convert2coord_pairs` call it from nopython mode over a whole ring, which is
where the line between the two modules falls.

The four `from timezonefinder.utils_numba import coord2int` sites follow, and
the three helpers join `utils.__all__`, which they had never been in despite
being reachable as `utils.coord2int` for as long as the aliases existed.
