+++ contributing/development/coding-design-and-state-management-rules.md
@@ -3,7 +3,7 @@
 - Define types centrally in `timezonefinder/configs.py` to avoid duplication and circular imports
 - **Do not let an unmeasured micro-optimisation choose the structure.** The shortcut index was first wired into `AbstractTimezoneFinder` as five flat array attributes with the slot arithmetic and entry decoding inlined at each call site, justified by an attribute hop "not worth the tidier grouping" — which was never measured. It costs nothing: routing the same lookup through `ShortcutIndex.entry_of` / `candidates_of` / `stop_index_of` measured inside the noise on all four strata, both estimators disagreeing. Encapsulate first, then measure, and only then trade the design away — with the numbers in the commit. A method call is not free, but at ~1 µs per query it is far below what this repository can resolve
 - **Removing the last caller of something is half the change: the other half is the callee.** Grep for remaining callers in the same commit, and act on what you find — no callers at all means delete it; callers only in tests means it is test scaffolding and should say so; callers only at build time means it must *leave the runtime modules*, because everything in `timezonefinder/utils_numba.py` is compiled and bound at import in every user's process. That is how `get_last_change_idx` outlived its query-path caller: precomputing it into the shortcut index removed the call, and the `@njit` function stayed behind being compiled for nobody. A helper that no longer runs per query is not merely tidy to move — it is cost every user pays
-- **A dispatch boundary costs more than any scalar per-query stage computes.** An empty `njit` call is ~98 ns and a cffi crossing the same order, against stages of 100-200 ns — so reaching for Numba or the C extension to speed one up is a measured dead end, and the [query-path decisions](../improvements/decisions/query-performance-and-shortcut-index-decisions.md) hold the numbers. **Applied in the other direction, 2026-09: `is_valid_lat`, `is_valid_lng` and `coord2int` are plain Python functions and must stay that way.** Each takes one number and does one operation, so `njit` cost about twice what the body did on every query. The rule is about the *decorator*, not the call — the fix is to remove `njit` from such a helper rather than to inline it at its call sites, which would buy a few more nanoseconds and leave the bound defined twice. `pt_in_poly_python` beside them is the contrast and the reason this is not an argument against Numba: an array of hundreds to tens of thousands of vertices, where the dispatch amortises to nothing. The compiled *vectorised* validators in `scripts/utils_numba.py` inline the same comparison instead of calling the scalar ones, because Numba cannot call a pure-Python function from nopython mode; `tests/test_property_validation.py` holds the two forms to each other.
+- **A dispatch boundary costs more than any scalar per-query stage computes.** An empty `njit` call is ~98 ns and a cffi crossing the same order, against stages of 100-200 ns — so reaching for Numba or the C extension to speed one up is a measured dead end, and the [query-path decisions](../improvements/decisions/query-performance-and-shortcut-index-decisions.md) hold the numbers. **Applied in the other direction, 2026-09: `is_valid_lat`, `is_valid_lng` and `coord2int` are plain Python functions in `timezonefinder/utils.py`, and must stay both.** Each takes one number and does one operation, so `njit` cost about twice what the body did on every query. The rule is about the *decorator*, not the call — the fix is to remove `njit` from such a helper rather than to inline it at its call sites, which would buy a few more nanoseconds and leave the bound defined twice. They live outside `utils_numba.py` on purpose: that module is what the package compiles, so a reader reaching for a function there is entitled to assume it is compiled, and an uncompiled one sitting among them is an invitation to re-add the decorator. `int2coord` stays inside it and stays compiled, which is where the line falls — `convert2coords` calls it from nopython mode over a whole ring. `pt_in_poly_python` is the same contrast at the other extreme, and the reason none of this is an argument against Numba: an array of hundreds to tens of thousands of vertices, where the dispatch amortises to nothing. The compiled *vectorised* validators in `scripts/utils_numba.py` inline the same comparison instead of calling the scalar ones, because Numba cannot call a pure-Python function from nopython mode; `tests/test_property_validation.py` holds the two forms to each other.
 - Preserve the fast lookup path. `prototypes/query_stage_profile.py` attributes a `timezone_at` query to its stages, per backend and per coordinate-access mode, off the committed fixtures — read its `FINDINGS` block before arguing about where query time goes, and re-run it rather than reasoning from a microbenchmark. **Those numbers are pinned to the commit they were taken at and go stale silently**: a change under `timezonefinder/` or to the packaged data can move every share quoted from them, so re-measuring belongs to the pull request that moves the critical path, not to a later one. They are also **one machine's**: rank on the counts a change removes (machine-independent) and on the `clang` / mapped column that a plain install actually runs, not on absolute nanoseconds. The [measurement baseline](../improvements/query-performance-measurement-baseline.md) holds the anchor, the denominators, the workload conversion and a one-command freshness check
 - Prefer dependency injection over module-level state; global helper functions are NOT thread-safe, concurrent workloads should use per-thread `TimezoneFinder(in_memory=True)` instances
 - **Declare each path/filename constant once** in the module that owns the resource and import it elsewhere; never re-derive a path or retype a filename string in a second file — the two copies drift when one is renamed
